# src/IO.Swagger/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "<h2> WILL BE UPDATED!</h2></br><b>Incomplete</b> API for car workshop system (lacks employees and parts management).</br><a href=\"https://app.swaggerhub.com/apis/soft_eng/project/0.1.1\">Swagger project (contains models)</a>.",
    "version" : "0.1.1",
    "title" : "Car Workshop",
    "contact" : {
      "email" : "k.baciejowski@gmail.com"
    }
  },
  "host" : "virtserver.swaggerhub.com",
  "basePath" : "/soft_eng/project/0.1.1",
...
  "schemes" : [ "https" ],
  "paths" : {
...
    "/visit" : {
      "post" : {
        "tags" : [ "visits" ],
        "summary" : "Schedule a visit",
        "operationId" : "addVisit",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "Information about scheduled visit. Required actions can be modified later on by an employee.",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/body_3"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Validation exception"
          }
        },
        "security" : [ {
          "workshop_auth" : [ "write:visits" ]
        } ]
      },
      "put" : {
        "tags" : [ "visits", "employees" ],
        "summary" : "Update a visit",
        "operationId" : "updateVisit",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Visit"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Validation exception"
          },
          "404" : {
            "description" : "Visit not found"
          }
        },
        "security" : [ {
          "workshop_auth" : [ "write:visits" ]
        } ]
      }
    },
    "/visit/find_by_status" : {
      "get" : {
        "tags" : [ "visits" ],
        "summary" : "Find visits by Status",
        "description" : "Multiple status values can be provided with comma separated strings",
        "operationId" : "findVisits",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "status",
          "in" : "query",
          "description" : "Status values that need to be considered for filter",
          "required" : true,
          "type" : "array",
          "items" : {
            "type" : "string",
            "default" : "planned",
            "enum" : [ "planned", "atService", "repaired", "checkedIn" ]
          },
          "collectionFormat" : "multi"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/inline_response_200"
              }
            }
          },
          "400" : {
            "description" : "Invalid status value"
          }
        },
        "security" : [ {
          "workshop_auth" : [ "read:visits" ]
        } ]
      }
    },
    "/visit/{visitId}" : {
      "get" : {
        "tags" : [ "visits" ],
        "summary" : "Find visit by Id",
        "operationId" : "getVisit",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "visitId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/inline_response_200"
              }
            }
          },
          "400" : {
            "description" : "Invalid id supplied"
          },
          "404" : {
            "description" : "Visit not found"
          }
        },
        "security" : [ {
          "workshop_auth" : [ "read:visits" ]
        } ]
      }
    },
    "/client/{clientId}/visits" : {
      "get" : {
        "tags" : [ "clients", "visits" ],
        "summary" : "Get all client's visits",
        "operationId" : "getClientVisits",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "clientId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Visit"
              }
            }
          },
          "404" : {
            "description" : "Client not found"
          }
        },
        "security" : [ {
          "workshop_auth" : [ "read:visits" ]
        } ]
      }
    },
...
    "/visit/{visitId}/maintain" : {
      "put" : {
        "tags" : [ "employees", "visits" ],
        "summary" : "Marks maintenance as finished",
        "operationId" : "maintain",
        "parameters" : [ {
          "name" : "visitId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Invalid visit id supplied"
          },
          "404" : {
            "description" : "Visit not found"
          }
        },
        "security" : [ {
          "workshop_auth" : [ "write:visits" ]
        } ]
      }
    },
    "/visit/{visitId}/diagnose" : {
      "put" : {
        "tags" : [ "employees", "visits" ],
        "summary" : "Marks diagnostic as finished",
        "operationId" : "diagnose",
        "parameters" : [ {
          "name" : "visitId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Invalid visit id supplied"
          },
          "404" : {
            "description" : "Visit not found"
          }
        },
        "security" : [ {
          "workshop_auth" : [ "write:visits" ]
        } ]
      }
    }
  },
  "securityDefinitions" : {
    "workshop_auth" : {
      "type" : "oauth2",
      "authorizationUrl" : "https://example.com/oauth/dialog",
      "flow" : "implicit",
      "scopes" : {
        "write:clients" : "modify clients",
        "read:clients" : "read clients",
        "write:cars" : "modify cars",
        "read:cars" : "read cars",
        "write:visits" : "modify visits",
        "read:visits" : "read visits",
        "write:payments" : "modify payments",
        "read:payments" : "read payments",
        "write:employees" : "modify employees",
        "read:employees" : "read employees"
      }
    },
    "api_key" : {
      "type" : "apiKey",
      "name" : "api_key",
      "in" : "header"
    }
  },
  "definitions" : {
    "Employee" : {
      "type" : "object",
      "required" : [ "employeeId", "name", "surname" ],
      "properties" : {
        "employeeId" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 13
        },
        "name" : {
          "type" : "string",
          "example" : "Katarzyna"
        },
        "surname" : {
          "type" : "string",
          "example" : "Nowak"
        }
      }
    },
...
    "Visit" : {
      "type" : "object",
      "required" : [ "date", "visitId" ],
      "properties" : {
        "visitId" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 5512
        },
        "date" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "2020-12-24T12:00:00+01"
        },
        "price" : {
          "type" : "number",
          "example" : 100.5
        },
        "requiredActions" : {
          "type" : "array",
          "example" : [ "oil change", "breaks checking" ],
          "items" : {
            "type" : "string"
          }
        },
        "status" : {
          "type" : "string",
          "example" : "atService",
          "enum" : [ "atService", "repaired", "checkedIn" ]
        },
        "assignedEmployees" : {
          "type" : "array",
          "example" : [ {
            "employeeId" : 12,
            "name" : "Bolesław",
            "surname" : "Nowak"
          }, {
            "employeeId" : 13,
            "name" : "Katarzyna",
            "surname" : "Nowak"
          } ],
          "items" : {
            "$ref" : "#/definitions/Employee"
          }
        }
      },
      "example" : {
        "date" : "2020-12-24T12:00:00+01",
        "requiredActions" : [ "oil change", "breaks checking" ],
        "visitId" : 5512,
        "price" : 100.5,
        "assignedEmployees" : [ {
          "employeeId" : 12,
          "name" : "Bolesław",
          "surname" : "Nowak"
        }, {
          "employeeId" : 13,
          "name" : "Katarzyna",
          "surname" : "Nowak"
        } ],
        "status" : "atService"
      }
    },
...
    "body_3" : {
      "type" : "object",
      "properties" : {
        "date" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "2020-12-24T12:00:00+01"
        },
        "requiredActions" : {
          "type" : "array",
          "items" : {
            "type" : "string",
            "example" : ""
          }
        }
      }
    },
    "inline_response_200" : {
      "type" : "object",
      "properties" : {
        "carId" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 1552
        },
        "visit" : {
          "$ref" : "#/definitions/Visit"
        }
      },
      "example" : {
        "visit" : {
          "date" : "2020-12-24T12:00:00+01",
          "requiredActions" : [ "oil change", "breaks checking" ],
          "visitId" : 5512,
          "price" : 100.5,
          "assignedEmployees" : [ {
            "employeeId" : 12,
            "name" : "Bolesław",
            "surname" : "Nowak"
          }, {
            "employeeId" : 13,
            "name" : "Katarzyna",
            "surname" : "Nowak"
          } ],
          "status" : "atService"
        },
        "carId" : 1552
      }
    }
  }
}